+++ packages/svelte-timeseries/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @qtsurfer/svelte-timeseries
 
+## 0.4.5
+
+### Patch Changes
+
+- Expose TimeSeriesFacade via onFacadeReady callback, add getDuckDB() and getChartBuilder() accessors
+
+- Updated dependencies []:
+  - @qtsurfer/sveltecharts@0.4.5
+
 ## 0.4.4
 
 ### Patch Changes